# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kasir Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Customers ─────────────────────────────────────────────
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: paymentMethod
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a new transaction (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionDetail"

  # ─── Dashboard ─────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, price, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        stock:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        stock:
          type: ["integer", "null"]
        isActive:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        stock:
          type: ["integer", "null"]
        isActive:
          type: boolean

    # ─── Customer ──────────────────────────────────────────
...
    Transaction:
      type: object
      required: [id, cashierName, paymentMethod, subtotal, tax, total, pointsEarned, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        cashierName:
          type: string
        paymentMethod:
          type: string
          enum: [cash, debit_card, credit_card, qris, transfer]
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
        amountPaid:
          type: number
        change:
          type: number
        pointsEarned:
          type: integer
        pointsUsed:
          type: integer
        status:
          type: string
          enum: [completed, voided]
        createdAt:
          type: string
    TransactionDetail:
      allOf:
        - $ref: "#/components/schemas/Transaction"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/TransactionItem"
    TransactionItem:
      type: object
      required: [id, productId, productName, quantity, price, subtotal]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
    TransactionInput:
      type: object
      required: [cashierName, paymentMethod, items]
      properties:
        customerId:
          type: ["integer", "null"]
        cashierName:
          type: string
        paymentMethod:
          type: string
          enum: [cash, debit_card, credit_card, qris, transfer]
        discount:
          type: number
        amountPaid:
          type: number
        pointsUsed:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/TransactionItemInput"
    TransactionItemInput:
      type: object
      required: [productId, quantity, price]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: number

    # ─── Dashboard ─────────────────────────────────────────
...