# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kasir Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  stock?: number | null;
  isActive?: boolean;
}