# lib/api-zod/src/generated/types/transactionDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kasir Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from './transaction';
import type { TransactionItem } from './transactionItem';

export type TransactionDetail = Transaction & {
  items: TransactionItem[];
};